refactor: split account-list-menu into account-menu re-usable component
