+++ ui/components/multichain/account-menu/account-menu.stories.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import { AccountMenu } from '.';
+
+export default {
+  title: 'Components/Multichain/AccountMenu',
+  component: AccountMenu,
+  argTypes: {
+    onClose: {
+      action: 'onClose',
+    },
+  },
+};
+
+export const DefaultStory = (args) => <AccountMenu {...args} />;
