+++ ui/components/multichain/account-menu/index.scss
@@ -1,0 +1,21 @@
+.multichain-account-menu-popover {
+  &__dialog {
+    max-height: 100%;
+  }
+
+  .popover-content {
+    border-radius: 0;
+  }
+
+  &__list {
+    overflow: auto;
+
+    &--menu-item-hidden {
+      opacity: 0.5;
+    }
+
+    &--menu-item-hidden-account {
+      opacity: 0.5;
+    }
+  }
+}
