+++ ui/components/multichain/account-menu/index.ts
@@ -1,0 +1,1 @@
+export { AccountMenu } from './account-menu';
